Shiva Unreal P1 and 2 transition condition

+++ scripts/Battle/Entity.cs
@@ -16,8 +16,8 @@ public abstract class Entity : MonoBehaviour, GotDamage
     [SerializeField]
     public int maxHP = 10;
     [SerializeField]
-    protected int currentHP;
-    public int healthPoint
+    protected float currentHP;
+    public float healthPoint
     {
         get
         {
